// lib/db/customer.ts
import { sql } from ".";
import { CustomerField, CustomersTableType } from "../types/customer";
import { formatCurrency } from "../utils";

export async function fetchCustomers() {
    try {
      const customers = await sql<CustomerField[]>`
        SELECT
          id,
          name
        FROM customers
        ORDER BY name ASC
      `;
  
      return customers;
    } catch (err) {
      console.error('Database Error:', err);
      throw new Error('Failed to fetch all customers.');
    }
}
  
export async function fetchFilteredCustomers(query: string) {
try {
    const data = await sql<CustomersTableType[]>`
        SELECT
        customers.id,
        customers.name,
        customers.email,
        customers.image_url,
        COUNT(invoices.id) AS total_invoices,
        SUM(CASE WHEN invoices.status = 'pending' THEN invoices.amount ELSE 0 END) AS total_pending,
        SUM(CASE WHEN invoices.status = 'paid' THEN invoices.amount ELSE 0 END) AS total_paid
        FROM customers
        LEFT JOIN invoices ON customers.id = invoices.customer_id
        WHERE
        customers.name ILIKE ${`%${query}%`} OR
        customers.email ILIKE ${`%${query}%`}
        GROUP BY customers.id, customers.name, customers.email, customers.image_url
        ORDER BY customers.name ASC
    `;

    const customers = data.map((customer) => ({
    ...customer,
    total_pending: formatCurrency(customer.total_pending),
    total_paid: formatCurrency(customer.total_paid),
    }));

    return customers;
} catch (err) {
    console.error('Database Error:', err);
    throw new Error('Failed to fetch customer table.');
}
}
  